+++ jsonjsdb-editor/test/fixtures/expected/__table__.json.js
@@ -1,0 +1,19 @@
+jsonjs.data['__table__'] = 
+[
+  {
+    "name": "evolution",
+    "last_modif": 1757694383
+  },
+  {
+    "name": "tag",
+    "last_modif": 1738189734
+  },
+  {
+    "name": "user",
+    "last_modif": 1738189737
+  },
+  {
+    "name": "__table__",
+    "last_modif": 1757694383
+  }
+]