+++ jsonjsdb-editor/test/fixtures/expected/tag.json.js
@@ -1,0 +1,18 @@
+jsonjs.data['tag'] = 
+[
+  {
+    "id": 1,
+    "name": "tag_1",
+    "parent_id": null
+  },
+  {
+    "id": 2,
+    "name": "tag_2",
+    "parent_id": 1
+  },
+  {
+    "id": 3,
+    "name": "tag_333",
+    "parent_id": 2
+  }
+]