+++ jsonjsdb_editor/test/excel_version/basic/convert_xlsxdb_to_jsonjsdb.js
@@ -1,8 +1,0 @@
-import { Jsonjsdb_editor, Jsonjsdb_watcher } from "../../../dist/index.js"
-
-// const editor = new Jsonjsdb_editor({readable: true})
-// await editor.set_output_db("app_db")
-// await editor.update_db("db")
-
-await Jsonjsdb_watcher.set_db("app_db")
-await Jsonjsdb_watcher.watch("db", true)